+++ src/components/layer-controller/RasterLayerController.js
@@ -4,6 +4,7 @@ import { getChannelStats, MAX_SLIDERS_AND_CHANNELS } from '@hms-dbmi/viv';
 import Grid from '@material-ui/core/Grid';
 import Button from '@material-ui/core/Button';
 import AddIcon from '@material-ui/icons/Add';
+import Slider from '@material-ui/core/Slider';
 
 import ExpansionPanel from '@material-ui/core/ExpansionPanel';
 import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary';
@@ -63,6 +64,7 @@ export default function RasterLayerController(props) {
   const { data, channels: channelOptions } = loader;
   const { labels, shape } = Array.isArray(data) ? data[data.length - 1] : data;
   const [domainType, setDomainType] = useState(layer.domainType);
+  const [isExpanded, setIsExpanded] = useState(true);
   const [globalLabelValues, setGlobalLabelValues] = useState(
     GLOBAL_LABELS
       .filter(field => typeof firstSelection[field] === 'number')
@@ -254,13 +256,34 @@ export default function RasterLayerController(props) {
   const classes = useExpansionPanelStyles();
   const summaryClasses = useExpansionPanelSummaryStyles();
   return (
-    <ExpansionPanel defaultExpanded className={classes.root}>
+    <ExpansionPanel
+      defaultExpanded
+      className={classes.root}
+      onChange={(e, expanded) => setIsExpanded(expanded)}
+      TransitionProps={{ enter: false }}
+    >
       <ExpansionPanelSummary
         expandIcon={<ExpandMoreIcon />}
         aria-controls={`layer-${name}-controls`}
         classes={{ ...summaryClasses }}
       >
-        {name}
+        <Grid container direction="column" m={1} justify="center">
+          <Grid item>{name}</Grid>
+          {!isExpanded && (
+            <Grid item>
+              <Slider
+                value={opacity}
+                onChange={(e, v) => setOpacity(v)}
+                valueLabelDisplay="auto"
+                getAriaLabel={() => 'opacity slider'}
+                min={0}
+                max={1}
+                step={0.01}
+                orientation="horizontal"
+              />
+            </Grid>
+          )}
+        </Grid>
       </ExpansionPanelSummary>
       <ExpansionPanelDetails className={classes.root}>
         <Grid item>
@@ -274,11 +297,7 @@ export default function RasterLayerController(props) {
             domainType={domainType}
             // Only allow for global dimension controllers that
             // exist in the `dimensions` part of the loader.
-            globalControlLabels={
-              labels.filter(
-                label => GLOBAL_LABELS.includes(label),
-              )
-            }
+            globalControlLabels={labels.filter(label => GLOBAL_LABELS.includes(label))}
             globalLabelValues={globalLabelValues}
             handleOpacityChange={setOpacity}
             handleColormapChange={setColormap}